add assistant api

+++ api/v1/bot.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Path, Body, Depends
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 
-from utils import log
+from utils import assistant, log
 from api.deps import db_client
 
 router = APIRouter()
@@ -25,13 +25,39 @@ class BotData(BaseModel):
     public: Optional[bool] = True
     created_at: Optional[int] = None
     updated_at: Optional[int] = None
+    # assistant_id: Optional[str] = None
+    model: Optional[str] = None
+    file_search: Optional[bool] = False
+    vector_store_ids: Optional[dict] = {}
+    code_interpreter: Optional[bool] = False
+    code_interpreter_files: Optional[dict]  = {}
+    functions: Optional[dict] = {}
+    temperature: Optional[float] = 1.0
 
 
 @router.post("/bot", name="add bot")
 async def bot_new(bot: BotData) -> Any:
     try:
+        assistant_id = None
+        if bot.model is None:
+            bot.model = "gpt-4o"
+        if bot.file_search or bot.code_interpreter or bot.functions:
+            new_assistant = assistant.create_assistant(
+                name=bot.name,
+                description=bot.description,
+                model=bot.model,
+                file_search=bot.file_search,
+                vector_store_ids=list(bot.vector_store_ids.keys()),
+                code_interpreter=bot.code_interpreter,
+                code_interpreter_files=list(bot.code_interpreter_files.values()),
+                functions=list(bot.functions.values()),
+                temperature=bot.temperature,
+            )
+            if new_assistant:
+                assistant_id = new_assistant.id
         created_at = int(time.time())
         updated_at = int(time.time())
+        print("backend bot:", bot)
         bot_id = db_client.bot.add_new_bot(
             name=bot.name,
             avatar=bot.avatar,
@@ -43,6 +69,14 @@ async def bot_new(bot: BotData) -> Any:
             public=bot.public,
             created_at=created_at,
             updated_at=updated_at,
+            assistant_id=assistant_id,
+            model=bot.model,
+            file_search=bot.file_search,
+            vector_store_ids=bot.vector_store_ids,
+            code_interpreter=bot.code_interpreter,
+            code_interpreter_files=bot.code_interpreter_files,
+            functions=bot.functions,
+            temperature=bot.temperature,
         )
     except Exception as err:
         log.debug(f"add bot error:{err}")
@@ -80,8 +114,16 @@ async def bot_info(bot_id: int) -> Any:
         "author_name": bot.author_name,
         "likes": bot.likes,
         "public": bot.public,
-        "created_at": created_at,
-        "updated_at": updated_at,
+        "created_at": bot.created_at,
+        "updated_at": bot.updated_at,
+        "assistant_id": bot.assistant_id,
+        "model": bot.model,
+        "file_search": bot.file_search,
+        "vector_store_ids": bot.vector_store_ids,
+        "code_interpreter": bot.code_interpreter,
+        "code_interpreter_files": bot.code_interpreter_files,
+        "functions": bot.functions,
+        "temperature": bot.temperature,
     }
     return JSONResponse(status_code=200, content=res)
 
@@ -109,6 +151,15 @@ async def bot_edit(bot_id: int, bot: BotData) -> Any:
             new_data["public"] = bot.public
             print("this is pub:", bot.public)
         new_data["updated_at"] = int(time.time())
+        new_data["model"] = bot.model
+        new_data["assistant_id"] = bot.assistant_id
+        new_data["file_search"] = bot.file_search
+        new_data["vector_store_ids"] = bot.vector_store_ids
+        new_data["code_interpreter"] = bot.code_interpreter
+        new_data["code_interpreter_files"] = bot.code_interpreter_files
+        new_data["functions"] = bot.functions
+        new_data["temperature"] = bot.temperature
+
         bot = db_client.bot.update_bot_by_id(
             bot_id,
             **new_data,
@@ -120,9 +171,15 @@ async def bot_edit(bot_id: int, bot: BotData) -> Any:
 
 
 @router.delete("/bot/{bot_id}", name="delete bot")
-async def bot_delete(bot_id: int) -> Any:
+async def bot_delete(bot_id: int, assistant_id: str) -> Any:
     try:
-        bot_id = db_client.bot.delete_bot(bot_id=bot_id)
+        del_asst = None
+        if assistant_id:
+            del_asst = assistant.delete_assistant(assistant_id)
+            if del_asst and del_asst.deleted == True:
+                db_client.bot.delete_bot(bot_id=bot_id)
+        else:
+            db_client.bot.delete_bot(bot_id=bot_id)
     except Exception as err:
         log.debug(f"delete bot error:{err}")
         return JSONResponse(status_code=500, content={"result": str(err)})
